refactor: move Navigation to nav-bar folder
