+++ nav-bar-sticky/src/components/nav-bar/Navigation.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import UnorderedList from './un-list/UnorderedList';
+import UnorderedList from './UnorderedList';
 
 const StyledNavigation = styled.nav``;
