+++ nav-bar-sticky/src/components/nav-bar/UnorderedList.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import { useAppDispatch } from '../../app/hooks';
 import { changer } from './pageSlice';
 import Item from './Item';
-import Link from '../Link';
+import Link from './Link';
 
 const StyledUnorderedList = styled.ul``;
